Use customer ephemeral key configure instead of checkout configure
