+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/model/PlaygroundCheckoutModel.kt
@@ -48,8 +48,6 @@ class CheckoutRequest private constructor(
     val paymentMethodOverrideRedisplay: AllowRedisplayFilter?,
     @SerialName("customer_session_payment_method_set_as_default")
     val paymentMethodSetAsDefaultFeature: FeatureState?,
-    @SerialName("customer_session_payment_method_sync_default")
-    val paymentMethodSyncDefaultFeature: FeatureState?,
 ) {
     @Serializable
     enum class CustomerKeyType {
@@ -77,7 +75,6 @@ class CheckoutRequest private constructor(
         private var paymentMethodRemoveFeature: FeatureState = FeatureState.Enabled
         private var paymentMethodRemoveLastFeature: FeatureState = FeatureState.Enabled
         private var paymentMethodSetAsDefaultFeature: FeatureState = FeatureState.Disabled
-        private var paymentMethodSyncDefaultFeature: FeatureState = FeatureState.Disabled
         private var paymentMethodRedisplayFeature: FeatureState = FeatureState.Enabled
         private var paymentMethodRedisplayFilters: List<AllowRedisplayFilter> = listOf(
             AllowRedisplayFilter.Unspecified,
@@ -146,10 +143,6 @@ class CheckoutRequest private constructor(
             this.paymentMethodSetAsDefaultFeature = state
         }
 
-        fun paymentMethodSyncDefaultFeature(state: FeatureState) {
-            this.paymentMethodSyncDefaultFeature = state
-        }
-
         fun paymentMethodRedisplayFeature(state: FeatureState) {
             this.paymentMethodRedisplayFeature = state
         }
@@ -184,7 +177,6 @@ class CheckoutRequest private constructor(
                 paymentMethodSaveFeature = paymentMethodSaveFeature,
                 paymentMethodRemoveFeature = paymentMethodRemoveFeature,
                 paymentMethodSetAsDefaultFeature = paymentMethodSetAsDefaultFeature,
-                paymentMethodSyncDefaultFeature = paymentMethodSyncDefaultFeature,
                 paymentMethodRemoveLastFeature = paymentMethodRemoveLastFeature,
                 paymentMethodRedisplayFeature = paymentMethodRedisplayFeature,
                 paymentMethodRedisplayFilters = paymentMethodRedisplayFilters,
