+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/model/PlaygroundCustomerModel.kt
@@ -21,7 +21,8 @@ class CustomerEphemeralKeyRequest private constructor(
     val paymentMethodRedisplayFeature: FeatureState?,
     @SerialName("customer_session_payment_method_allow_redisplay_filters")
     val paymentMethodRedisplayFilters: List<AllowRedisplayFilter>?,
-
+    @SerialName("customer_session_payment_method_sync_default")
+    val paymentMethodSyncDefaultFeature: FeatureState?,
 ) {
     @Serializable
     enum class CustomerKeyType {
@@ -38,6 +39,7 @@ class CustomerEphemeralKeyRequest private constructor(
         private var merchantCountryCode: String? = null
         private var paymentMethodRemoveFeature: FeatureState = FeatureState.Enabled
         private var paymentMethodRemoveLastFeature: FeatureState = FeatureState.Enabled
+        private var paymentMethodSyncDefaultFeature: FeatureState = FeatureState.Disabled
         private var paymentMethodRedisplayFilters: List<AllowRedisplayFilter> = listOf(
             AllowRedisplayFilter.Unspecified,
             AllowRedisplayFilter.Limited,
@@ -60,6 +62,10 @@ class CustomerEphemeralKeyRequest private constructor(
             this.paymentMethodRemoveLastFeature = state
         }
 
+        fun paymentMethodSyncDefaultFeature(state: FeatureState) {
+            this.paymentMethodSyncDefaultFeature = state
+        }
+
         fun paymentMethodRedisplayFilters(filters: List<AllowRedisplayFilter>) {
             this.paymentMethodRedisplayFilters = filters
         }
@@ -76,6 +82,7 @@ class CustomerEphemeralKeyRequest private constructor(
                 paymentMethodSaveFeature = FeatureState.Enabled,
                 paymentMethodRemoveFeature = paymentMethodRemoveFeature,
                 paymentMethodRemoveLastFeature = paymentMethodRemoveLastFeature,
+                paymentMethodSyncDefaultFeature = paymentMethodSyncDefaultFeature,
                 paymentMethodRedisplayFeature = FeatureState.Enabled,
                 paymentMethodRedisplayFilters = paymentMethodRedisplayFilters,
             )
